+++ worker/src/admin_api/webhook_settings.ts
@@ -1,7 +1,7 @@
 import { Context } from "hono";
 import { HonoCustomType } from "../types";
 import { CONSTANTS } from "../constants";
-import { AdminWebhookSettings } from "../models/models";
+import { AdminWebhookSettings } from "../models";
 
 async function getWebhookSettings(c: Context<HonoCustomType>): Promise<Response> {
     const settings = await c.env.KV.get<AdminWebhookSettings>(CONSTANTS.WEBHOOK_KV_SETTINGS_KEY, "json");