+++ worker/src/telegram_api/settings.ts
@@ -6,17 +6,24 @@ export class TelegramSettings {
     enableAllowList: boolean;
     allowList: string[];
     miniAppUrl: string;
+    enableGlobalMailPush: boolean;
+    globalMailPushList: string[];
 
-    constructor(enableAllowList: boolean, allowList: string[], miniAppUrl: string) {
+    constructor(
+        enableAllowList: boolean, allowList: string[], miniAppUrl: string,
+        enableGlobalMailPush: boolean, globalMailPushList: string[]
+    ) {
         this.enableAllowList = enableAllowList;
         this.allowList = allowList;
         this.miniAppUrl = miniAppUrl;
+        this.enableGlobalMailPush = enableGlobalMailPush;
+        this.globalMailPushList = globalMailPushList;
     }
 }
 
 async function getTelegramSettings(c: Context<HonoCustomType>): Promise<Response> {
     const settings = await c.env.KV.get<TelegramSettings>(CONSTANTS.TG_KV_SETTINGS_KEY, "json");
-    return c.json(settings || new TelegramSettings(false, [], ""));
+    return c.json(settings || new TelegramSettings(false, [], "", false, []));
 }