+++ worker/src/user_api/settings.ts
@@ -1,9 +1,12 @@
+import { Context } from "hono";
+
+import { HonoCustomType } from "../types";
 import { UserSettings } from "../models";
 import { getJsonSetting } from "../utils"
 import { CONSTANTS } from "../constants";
 
 export default {
-    openSettings: async (c) => {
+    openSettings: async (c: Context<HonoCustomType>) => {
         const value = await getJsonSetting(c, CONSTANTS.USER_SETTINGS_KEY);
         const settings = new UserSettings(value);
         return c.json({
@@ -11,7 +14,7 @@ export default {
             enableMailVerify: settings.enableMailVerify,
         })
     },
-    settings: async (c) => {
+    settings: async (c: Context<HonoCustomType>) => {
         const user = c.get("userPayload");
         // check if user exists
         const db_user_id = await c.env.DB.prepare(